feat: SNPureCrypto implementation with libsodium + cryptojs
