+++ package.json
@@ -26,6 +26,7 @@
     "@types/crypto-js": "^4.0.1",
     "@types/jest": "^26.0.15",
     "@types/jsdom": "^16.2.5",
+    "@types/libsodium-wrappers": "^0.7.8",
     "@types/node": "^14.14.6",
     "@typescript-eslint/eslint-plugin": "^4.6.1",
     "@typescript-eslint/parser": "^4.6.1",
@@ -35,8 +36,8 @@
     "eslint": "^7.13.0",
     "jest": "^26.6.3",
     "jsdom": "^16.4.0",
+    "libsodium-wrappers": "^0.7.8",
     "nock": "^13.0.5",
-    "sodium-plus": "^0.9.0",
     "ts-loader": "^8.0.9",
     "tslint": "^6.1.3",
     "typedoc": "^0.19.2",
